requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
certifi==2020.6.20        # via requests
cffi==1.14.0              # via cryptography
chardet==3.0.4            # via requests
click==7.1.2              # via pip-tools
cryptography==2.9.2       # via jwcrypto
github3.py==1.3.0         # via -r requirements.in
idna==2.9                 # via requests
jwcrypto==0.7             # via github3.py
numpy==1.19.0             # via pandas
pandas==1.0.5             # via -r requirements.in
pillow==7.1.2             # via -r requirements.in, pytesseract
pip-tools==5.2.1          # via -r requirements.in
pycparser==2.20           # via cffi
pytesseract==0.3.4        # via -r requirements.in
python-dateutil==2.8.1    # via github3.py, pandas
pytz==2020.1              # via pandas
requests==2.24.0          # via github3.py
six==1.15.0               # via cryptography, pip-tools, python-dateutil
uritemplate==3.0.1        # via github3.py
urllib3==1.25.9           # via requests

# The following packages are considered to be unsafe in a requirements file:
# pip
